tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/size.attributes.parse.exp.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.parse.exp</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>

<h1>size.attributes.parse.exp</h1>
<p class="desc">Parsing of non-negative integers</p>


<script>
var t = async_test("Parsing of non-negative integers");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = '100e1';
offscreenCanvas.height = '100e1';
_assertSame(offscreenCanvas.width, 1000.0, "offscreenCanvas.width", "1000.0");
_assertSame(offscreenCanvas.height, 1000.0, "offscreenCanvas.height", "1000.0");
t.done();

});
</script>
